#25855: Add DP Support and move Yolov12x to models/demos
