+++ .github/CODEOWNERS
@@ -273,7 +273,7 @@ models/experimental/classification_eval @tenstorrent/cse-developer-ttnn
 models/demos/yolov11 @ssinghalTT @tenstorrent/cse-developer-ttnn
 models/experimental/segmentation_evaluation @tenstorrent/cse-developer-ttnn
 models/demos/yolov6l @ssinghalTT @tenstorrent/cse-developer-ttnn
-models/experimental/yolov12x @ssinghalTT @tenstorrent/cse-developer-ttnn
+models/demos/yolov12x @ssinghalTT @tenstorrent/cse-developer-ttnn
 models/experimental/efficientnetb0 @ssinghalTT @tenstorrent/cse-developer-ttnn
 models/experimental/yolov5x @ssinghalTT @tenstorrent/cse-developer-ttnn
 models/experimental/vovnet @ssinghalTT @tenstorrent/cse-developer-ttnn
