+++ .github/workflows/perf-device-models-impl.yaml
@@ -317,7 +317,7 @@ jobs:
         with:
           model_name: "yolov12x"
           commands: |
-            pytest models/experimental/yolov12x/tests/perf -m models_device_performance_bare_metal
+            pytest models/demos/yolov12x/tests/perf -m models_device_performance_bare_metal
         if: ${{ !cancelled() && matrix.test-info.civ2-compatible && !matrix.test-info.job-set1 }}
         timeout-minutes: ${{ matrix.test-info.timeout }}
