+++ models/demos/yolov12x/demo/images/bus.jpg
