--- models/experimental/yolov12x/tests/pcc/test_ttnn_yolov12x_c3k.py
@@ -1,71 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC.
-
-# SPDX-License-Identifier: Apache-2.0
-
-import ttnn
-import pytest
-
-from tests.ttnn.utils_for_testing import assert_with_pcc
-
-from models.experimental.yolov12x.reference import yolov12x
-from models.experimental.yolov12x.tt.c3k2 import TtnnC3k
-from models.experimental.yolov12x.tt.model_preprocessing import (
-    create_yolov12x_input_tensors_submodules,
-    create_yolov12x_model_parameters,
-)
-from models.experimental.yolov12x.common import YOLOV12_L1_SMALL_SIZE
-
-
-@pytest.mark.parametrize(
-    "in_channel, out_channel, fwd_input_shape",
-    [
-        ([96, 96, 96, 48, 48, 48, 48], [48, 48, 96, 48, 48, 48, 48], [1, 96, 160, 160]),  # 1
-        ([192, 192, 192, 96, 96, 96, 96], [96, 96, 192, 96, 96, 96, 96], [1, 192, 80, 80]),  # 2
-        ([384, 384, 384, 192, 192, 192, 192], [192, 192, 384, 192, 192, 192, 192], [1, 384, 40, 40]),  # 3
-        ([384, 384, 384, 192, 192, 192, 192], [192, 192, 384, 192, 192, 192, 192], [1, 384, 20, 20]),  # 4
-    ],
-)
-@pytest.mark.parametrize(
-    "kernel, stride, padding, dilation, groups",
-    [
-        (
-            [1, 1, 1, 3, 3, 3, 3],
-            [1, 1, 1, 1, 1, 1, 1],
-            [0, 0, 0, 1, 1, 1, 1],
-            [1, 1, 1, 1, 1, 1, 1],
-            [1, 1, 1, 1, 1, 1, 1],
-        ),
-    ],
-)
-@pytest.mark.parametrize("device_params", [{"l1_small_size": YOLOV12_L1_SMALL_SIZE}], indirect=True)
-def test_yolov12x_c3k(
-    device,
-    reset_seeds,
-    in_channel,
-    out_channel,
-    fwd_input_shape,
-    kernel,
-    stride,
-    padding,
-    dilation,
-    groups,
-):
-    torch_module = yolov12x.C3k(in_channel, out_channel, kernel, stride, padding, dilation, groups)
-    torch_module.eval()
-    torch_input, ttnn_input = create_yolov12x_input_tensors_submodules(
-        device,
-        batch_size=fwd_input_shape[0],
-        input_channels=fwd_input_shape[1],
-        input_height=fwd_input_shape[2],
-        input_width=fwd_input_shape[3],
-    )
-    ttnn_input = ttnn.to_device(ttnn_input, device=device)
-    ttnn_input = ttnn.to_layout(ttnn_input, layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
-    torch_output = torch_module(torch_input)
-    parameters = create_yolov12x_model_parameters(torch_module, torch_input, device=device)
-    ttnn_module = TtnnC3k(device=device, parameter=parameters.conv_args, conv_pt=parameters)
-    ttnn_output = ttnn_module(ttnn_input)
-    ttnn_output = ttnn.to_torch(ttnn_output)
-    ttnn_output = ttnn_output.permute(0, 3, 1, 2)
-    ttnn_output = ttnn_output.reshape(torch_output.shape)
-    assert_with_pcc(torch_output, ttnn_output, 0.99)
