--- models/experimental/yolov12x/tests/perf/test_e2e_performant.py
@@ -1,59 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-
-import time
-import pytest
-from loguru import logger
-import ttnn
-from models.experimental.yolov12x.runner.performant_runner import YOLOv12xPerformantRunner
-from models.utility_functions import run_for_wormhole_b0
-from models.experimental.yolov12x.common import YOLOV12_L1_SMALL_SIZE
-
-
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype",
-    ((1, ttnn.bfloat8_b, ttnn.bfloat8_b),),
-)
-@pytest.mark.parametrize(
-    "resolution",
-    [
-        (640, 640),
-    ],
-)
-@pytest.mark.models_performance_bare_metal
-@run_for_wormhole_b0()
-@pytest.mark.parametrize(
-    "device_params",
-    [{"l1_small_size": YOLOV12_L1_SMALL_SIZE, "trace_region_size": 6434816, "num_command_queues": 2}],
-    indirect=True,
-)
-def test_e2e_performant(
-    device,
-    batch_size,
-    act_dtype,
-    weight_dtype,
-    model_location_generator,
-    resolution,
-):
-    performant_runner = YOLOv12xPerformantRunner(
-        device,
-        batch_size,
-        act_dtype,
-        weight_dtype,
-        resolution=resolution,
-        model_location_generator=model_location_generator,
-    )
-
-    t0 = time.time()
-    for _ in range(10):
-        _ = performant_runner.run()
-    ttnn.synchronize_device(device)
-    t1 = time.time()
-
-    performant_runner.release()
-
-    inference_time_avg = round((t1 - t0) / 10, 6)
-    logger.info(
-        f"ttnn_yolov12x_batch_size: {batch_size}, resolution: {resolution}. One inference iteration time (sec): {inference_time_avg}, FPS: {round(batch_size/inference_time_avg)}"
-    )
