+++ tests/nightly/single_card/yolov12x/test_ttnn_yolov12x.py
@@ -1,1 +1,1 @@
-../../../../models/experimental/yolov12x/tests/pcc/test_ttnn_yolov12x.py
+../../../../models/demos/yolov12x/tests/pcc/test_ttnn_yolov12x.py
