+++ tests/scripts/single_card/run_single_card_demo_tests.sh
@@ -338,7 +338,7 @@ run_yolov6l_demo() {
 
 run_yolov12x_demo() {
 
-  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto --disable-warnings  models/experimental/yolov12x/demo/demo.py --timeout 600
+  pytest -n auto --disable-warnings  models/demos/yolov12x/demo/demo.py --timeout 600
 
 }
